second commit

+++ app/api/item/[id]/route.ts
@@ -1,9 +1,14 @@
 import { NextRequest, NextResponse } from 'next/server';
-import  prisma  from '../../../lib/prisma'
+import prisma from '../../../lib/prisma';
 
-export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
-  const id = parseInt(params.id);
+export async function PUT(req: NextRequest) {
+  const url = new URL(req.url);
+  const id = parseInt(url.pathname.split('/').pop() || '', 10);
   const data = await req.json();
+
+  if (isNaN(id)) {
+    return new NextResponse('Invalid item ID', { status: 400 });
+  }
 
   const updatedItem = await prisma.item.update({
     where: { id },
@@ -21,8 +26,13 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
   return NextResponse.json(updatedItem);
 }
 
-export async function DELETE(_: NextRequest, { params }: { params: { id: string } }) {
-  const id = parseInt(params.id);
+export async function DELETE(req: NextRequest) {
+  const url = new URL(req.url);
+  const id = parseInt(url.pathname.split('/').pop() || '', 10);
+
+  if (isNaN(id)) {
+    return new NextResponse('Invalid item ID', { status: 400 });
+  }
 
   await prisma.item.delete({
     where: { id },
